+++ src/components/Cart.jsx
@@ -1,0 +1,8 @@
+/* Rename component
+ * This is used for template and routes
+ */
+const Cart = () => {
+  return <section id="cart">Cart</section>;
+};
+
+export default Cart;