+++ src/components/Loading.jsx
@@ -1,0 +1,11 @@
+import LoadingIcon from "../assets/icons/progress_activity.svg?react";
+
+const Loading = () => {
+  return (
+    <div>
+      <LoadingIcon />
+    </div>
+  );
+};
+
+export default Loading;