+++ src/components/Placeholder_01.jsx
@@ -1,8 +1,0 @@
-/* Rename component
- * This is used for template and routes
- */
-const Placeholder_01 = () => {
-  return <section id="placeholder_01">Placeholder_01</section>;
-};
-
-export default Placeholder_01;